Refactor item change notification to fix observers
In order to keep proper accounting of observers added and removed from
the `AVPlayerItem` in the current player, the
`mediaPlayerCurrentItemDidChange:` method that is called on a player
item change now utilizes the change dictionary to get the old and new
items and register and deregister notifications appropriately. In order
for this to also work on player change notifications, a dummy change
dictionary is constructed and passed along.