Fix for asset mismatch when setting player
If a player delegate substitutes a different asset than the one arriving
in the `playerItem`, then we should use that asset when determining what
to do about playback during `setPlayerItem`. This change calls
`playerItemAsset`, which allows the delegate to make a substitution,
before making comparisons with the previous player asset.